chore: 스타일 컴포넌트 스토리북 설정
