+++ frontend/.storybook/preview.tsx
@@ -1,0 +1,33 @@
+import React from 'react';
+import type { Preview } from '@storybook/react';
+import { ThemeProvider } from 'styled-components';
+import { withThemeFromJSXProvider } from '@storybook/addon-styling';
+
+import GlobalStyles from '../src/styles/globalStyle';
+import { lightTheme, darkTheme } from '../src/styles/theme';
+
+const preview: Preview = {
+  parameters: {
+    actions: { argTypesRegex: '^on[A-Z].*' },
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/,
+      },
+    },
+  },
+};
+
+export const decorators = [
+  withThemeFromJSXProvider({
+    themes: {
+      light: lightTheme,
+      dark: darkTheme,
+    },
+    defaultTheme: 'light',
+    Provider: ThemeProvider,
+    GlobalStyles,
+  }),
+];
+
+export default preview;
