+++ frontend/package.json
@@ -25,6 +25,7 @@
     "@storybook/addon-essentials": "^7.0.25",
     "@storybook/addon-interactions": "^7.0.25",
     "@storybook/addon-links": "^7.0.25",
+    "@storybook/addon-styling": "^1.3.2",
     "@storybook/blocks": "^7.0.25",
     "@storybook/react": "^7.0.25",
     "@storybook/react-webpack5": "^7.0.25",
